Move make_temp, mkdir, set_dir_permission from fs.rs to ops/fs.rs
